requirements.txt
aleph-alpha-client
langchain
qdrant-client
python-dotenv
scipy
sklearn
streamlit